Adding SecureKeychainTokenStore
